Refactor code and break it to individual components
